litesd: Add a litesdcard model
This change adds a simple model for the LiteX litesdcard gateware, for
SD/MMC support in microwatt.

The litesd core has a set of banks of registers for the phy, core, dma
(reader & writer) and irq blocks. We keep these as separate blocks in
case we need to re-arrange in future.

CC: Philippe Mathieu-Daudé <f4bug@amsat.org>
Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>